Aberrations inlcuded
spreadsheet now calculates for the four sample Dall-Kirkham examples the net spherical aberration at the focus AND the coma at one field angle. The coma estimate is based on an example detector with 11-micron pixels and uses a radial position 1/3 of the way out along the diagonal of the focal plane array. 

Peak-to-valley and RMS values of spherical aberration and coma are provided in the spreadsheet.